requirements.txt
# YouTube Transcript API
youtube-transcript-api==0.6.1

# Wikipedia API for Wikification
wikipedia-api==0.5.8
wikimapper==0.1.5

# Topic Modeling
scikit-learn==1.3.2
gensim==4.3.2
nltk==3.8.1

# Data Processing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4

# Web Framework (optional, for API)
fastapi==0.104.1
uvicorn==0.24.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0 